Allow to set the API version to use separately by endpoint
After https://issues.redhat.com/browse/SWATCH-2410, we need to increase the version of one endpoint to avoid introducing breaking changes.

To do the above, we need to make these changes, so we can set the version by each endpoint.
