+++ api/rhsm-subscriptions-api-spec.yaml
@@ -5,13 +5,13 @@ info:
   version: 1.0.0
 
 servers:
-  - url: /{PATH_PREFIX}/{APP_NAME}/v1
+  - url: /{PATH_PREFIX}/{APP_NAME}
     variables:
       PATH_PREFIX:
         default: api
       APP_NAME:
         default: rhsm-subscriptions
-  - url: https://{HOSTNAME}/{PATH_PREFIX}/{APP_NAME}/v1
+  - url: https://{HOSTNAME}/{PATH_PREFIX}/{APP_NAME}
     variables:
       HOSTNAME:
         enum:
@@ -26,7 +26,7 @@ servers:
         default: rhsm-subscriptions
 
 paths:
-  /version:
+  /v1/version:
     description: "Provides version and build information about the deployed application."
     get:
       summary: "Request version and build information about the deployed application."
@@ -42,11 +42,13 @@ paths:
                 $ref: "#/components/schemas/VersionInfo"
         '500':
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
-  /opt-in:
+  /v1/opt-in:
     description: "Endpoint for opting into subscription watch."
     get:
       summary: "Get the opt-in configuration for the current account."
       operationId: getOptInConfig
+      tags:
+        - opt-in
       responses:
         '200':
           description: "The request for the account's opt-in configuration was successful."
@@ -76,6 +78,8 @@ paths:
       summary: "Create/Update an account's opt-in configuration. Org ID are defined by the
                 identity header."
       operationId: putOptInConfig
+      tags:
+        - opt-in
       responses:
         '200':
           description: "The request for the account's opt-in configuration was successful."
@@ -105,6 +109,8 @@ paths:
       summary: "Delete an opt-in config for the associated org. The org ID are
                 pulled from the identity header."
       operationId: deleteOptInConfig
+      tags:
+        - opt-in
       responses:
         '204':
           description: "Successfully deleted the opt-in configuration for the accociated org"
@@ -116,7 +122,7 @@ paths:
           $ref: "../spec/error-responses.yaml#/$defs/ResourceNotFound"
         '500':
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
-  /tally/products/{product_id}/{metric_id}:
+  /v1/tally/products/{product_id}/{metric_id}:
     description: "Operations on a data set of a given tally report."
     parameters:
       - name: product_id
@@ -234,7 +240,7 @@ paths:
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
       tags:
         - tally
-  /capacity/products/{product_id}/{metric_id}:
+  /v1/capacity/products/{product_id}/{metric_id}:
     description: 'Operations for capacity report for a given account and product'
     parameters:
       - name: product_id
@@ -345,7 +351,7 @@ paths:
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
       tags:
         - capacity
-  /instances/{id}/guests:
+  /v1/instances/{id}/guests:
     description: 'Operations for instance-id mappings.'
     parameters:
       - name: id
@@ -403,7 +409,7 @@ paths:
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
       tags:
         - instances
-  /instances/products/{product_id}:
+  /v1/instances/products/{product_id}:
     parameters:
       - name: product_id
         description: The ID for the product we wish to query
@@ -507,7 +513,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/InstanceResponse'
-  /subscriptions/products/{product_id}:
+  /v1/subscriptions/products/{product_id}:
     description: "Operations for total capacity by SKU for all of the org's active subscriptions for given Swatch product ID."
     parameters:
       - name: product_id
@@ -531,6 +537,7 @@ paths:
     get:
       summary: "Returns the total capacity by SKU for all of the org's active subscriptions for given Swatch product ID."
       operationId: getSkuCapacityReport
+      deprecated: true
       parameters:
         - name: category
           in: query
@@ -1267,7 +1274,6 @@ components:
           $ref: "#/components/schemas/Uom"
         metric_id:
           type: string
-
     SkuCapacityReportSort:
       type: string
       enum:
