+++ src/main/java/org/candlepin/subscriptions/resource/api/InternalTallyApiResource.java
@@ -20,18 +20,23 @@
  */
 package org.candlepin.subscriptions.resource.api;
 
-import org.candlepin.subscriptions.utilization.admin.api.InternalSubscriptionSyncOpenapiJsonApi;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
+import org.candlepin.subscriptions.tally.admin.api.RootApi;
 import org.springframework.stereotype.Component;
 
-/** Serves the OpenAPI json for the internal subscription sync API */
+/** Serves the OpenAPI json for the internal tally API */
 @Component
-public class InternalSubscriptionSyncApiJsonResource
-    implements InternalSubscriptionSyncOpenapiJsonApi {
-  @Autowired ApiSpecController controller;
+@AllArgsConstructor
+public class InternalTallyApiResource implements RootApi {
+  private final ApiSpecController controller;
 
   @Override
   public String getOpenApiJson() {
-    return controller.getInternalSubSyncApiJson();
+    return controller.getInternalTallyApiJson();
+  }
+
+  @Override
+  public String getOpenApiYaml() {
+    return controller.getInternalTallyApiYaml();
   }
 }
